Resource world market rework (WIP)
