+++ src/main/java/me/pugabyte/nexus/features/shops/providers/_ShopProvider.java
@@ -10,7 +10,6 @@ import lombok.Getter;
 import me.pugabyte.nexus.features.menus.MenuUtils;
 import me.pugabyte.nexus.features.shops.Shops;
 import me.pugabyte.nexus.features.shops.providers.BrowseProductsProvider.ShulkerContentsProvider;
-import me.pugabyte.nexus.framework.exceptions.postconfigured.InvalidInputException;
 import me.pugabyte.nexus.models.banker.BankerService;
 import me.pugabyte.nexus.models.shop.Shop.Product;
 import me.pugabyte.nexus.models.shop.Shop.ShopGroup;
@@ -53,11 +52,6 @@ public abstract class _ShopProvider extends MenuUtils implements InventoryProvid
 		this.page = page;
 		this.shopGroup = ShopGroup.of(viewer);
 		try {
-			if (!(this instanceof BrowseMarketProvider || this instanceof SearchProductsProvider) && viewer.getWorld().getName().startsWith("resource"))
-				throw new InvalidInputException("You cannot use player shops while in the resource world");
-			if ((this instanceof YourShopProvider) && !viewer.hasPermission("shops.edit"))
-				throw new InvalidInputException("Shops is currently in beta testing, only Trusted and above can create shops");
-
 			SmartInventory.builder()
 					.provider(provider)
 					.title(colorize(title))
