+++ src/main/java/me/pugabyte/nexus/models/boost/Boostable.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
-import me.pugabyte.nexus.features.shops.Shops.Market;
+import me.pugabyte.nexus.features.shops.Market;
 import me.pugabyte.nexus.utils.ItemBuilder;
 import org.bukkit.Material;
 import org.jetbrains.annotations.NotNull;
